chia|tests: Move some stuff from `plot_tools.py` into new file `util.py`
