+++ chia/harvester/harvester.py
@@ -7,7 +7,8 @@ from typing import Callable, Dict, List, Optional, Tuple
 
 import chia.server.ws_connection as ws  # lgtm [py/import-and-import-from]
 from chia.consensus.constants import ConsensusConstants
-from chia.plotting.plot_tools import PlotsRefreshParameter, PlotManager, PlotRefreshResult
+from chia.plotting.plot_tools import PlotManager
+from chia.plotting.util import PlotsRefreshParameter, PlotRefreshResult
 
 log = logging.getLogger(__name__)
